Verification - fix

+++ model_shared/src/commonMain/kotlin/com/nowiwr01p/model/work/timer/TimerType.kt
@@ -1,6 +1,6 @@
 package com.nowiwr01p.model.work.timer
 
-sealed class TimerType(val value: Long) {
-    data class Up(val endValue: Long): TimerType(endValue)
-    data class Down(val startValue: Long): TimerType(startValue)
+sealed class TimerType(val value: Int) {
+    data class Up(val endValue: Int): TimerType(endValue)
+    data class Down(val startValue: Int): TimerType(startValue)
 }
